refactor from switch to cliffy
