+++ plugins/platform-backend/cli/cli.ts
@@ -1,4 +1,4 @@
-import { parse, path, yaml, Entity } from "./deps.ts";
+import { parse, path, yaml, Entity, Command } from "./deps.ts";
 
 const usage = (name: string, description: string) => `
 ${name}: ${description}
@@ -25,53 +25,51 @@ export async function cli(options: CLIOptions) {
   let flags = parse(args);
   let [command] = flags._;
 
-  try {
-    switch (command) {
-      case "info": {
-        let ref = await findEntityContext(flags);
-        let response = await get(`components/${ref}/info`);
-        if (response.ok) {
-          await Deno.stdout.write(new TextEncoder().encode(await response.text()));
+  const cmd = new Command()
+    .name(name)
+    .version(() => `${name}\n${Array(name.length).fill("=").join('')}\narchitecture: ${target}\nbackstage: ${apiURL}`)
+    .description(description)
+    .command('info', 'display info about a backstage component entity.')
+    .option('-c --component <component:string>', 'The backstage component entity')
+    .action(async ({ component }) => {
+      let ref = await findEntityContext(component);
+      let response = await get(`components/${ref}/info`);
+      if (response.ok) {
+        await Deno.stdout.write(new TextEncoder().encode(await response.text()));
+      } else {
+        if (response.status === 404) {
+          throw new MainError(`unknown component '${ref}'`);
         } else {
-          if (response.status === 404) {
-            throw new MainError(`unknown component '${ref}'`);
-          } else {
-            throw new MainError(`communication error with backstage server: ${response.status} ${response.statusText}`);
-          }
+          throw new MainError(`communication error with backstage server: ${response.status} ${response.statusText}`);
         }
-        break;
       }
-      case "version":
-        console.log(`${name}\n${Array(name.length).fill("=").join('')}\narchitecture: ${target}\nbackstage: ${apiURL}`);
-        break;
-      case "help":
-      default:
-        console.log(usage(name, description))
-        break;
+    })
+
+    try {
+      await cmd.parse(args);
+    } catch (error) {
+      if (error instanceof MainError) {
+        console.log(error.message);
+      } else {
+        throw error;
+      }
     }
-  } catch (error) {
-    if (error instanceof MainError) {
-      console.log(error.message);
-    } else {
-      throw error;
-    }
-  }
 }
 
-async function findEntityContext(flags: ReturnType<typeof parse>): Promise<string> {
-  if (flags.component) {
-    return flags.component;
-  } else {
-    let catalogInfoYaml = await findAndRead("catalog-info.yml", "catalog-info.yaml");
-    if (catalogInfoYaml.found) {
-      let [info] = yaml.parseAll(catalogInfoYaml.content) as Iterable<Entity>;
-      if (info && info.metadata?.name) {
-        return info.metadata.name;
-      }
-    } else {
-      throw new MainError('unable to determine the component. You can set it explicitly by passing the `--component` flag');
-    }
+async function findEntityContext(component?: string): Promise<string> {
+  if (component) {
+    return component;
   }
+  let catalogInfoYaml = await findAndRead("catalog-info.yml", "catalog-info.yaml");
+  if (catalogInfoYaml.found) {
+    let [info] = yaml.parseAll(catalogInfoYaml.content) as Iterable<Entity>;
+    if (info && info.metadata?.name) {
+      return info.metadata.name;
+    }
+  } else {
+    throw new MainError('unable to determine the component. You can set it explicitly by passing the `--component` flag');
+  }
+
   return '';
 }
