+++ plugins/platform-backend/cli/deno.json
@@ -1,4 +1,7 @@
 {
+  "tasks": {
+    "dev": "deno run --allow-read --allow-net=localhost:7007 tasks/dev.ts"
+  },
   "lint": {
     "rules": {
       "exclude": ["prefer-const"]
